Robotic Hand Control using Human Hand Gesture recognition
It is an image processing based project. MATLAB was used to process the
live video and Arduino controller was used to control the human hand
small replica. This model had five fingers same as that of human hand
and bent in the same way as that of human fingers. Every change in the
movement of human finger in front of the camera is replicated in the
mechanical hand model. This project interfaced MATLAB with controller.
MATLAB processed real time hand movement and simultaneously sends
command to the controller.